removed bs4 dependency and fixed 2 bugs introduced during earlier refactoring
there were to functions named fetch_data.  the first was not used and
depended on beautiful soup, because initial attempts were to use bs4
to parse the HTML.  that's been removed and gets rid of the need to
import bs4.  this also reduces load time.

the other bug was in the second, actually used version of fetch_data.
it used to be part of main, and returned an integer when it
encountered an error because main returns the integer that is the
intended exit status.  when this was refactored to return a dictionary
of scraped variables, these should have been turned into sys.exit
calls (or maybe defined my own exceptions); when these errors
occurred, the integer ended up being used as a dictionary.  this is
now fixed.