tools.
compress-coupon-csv gets rid of useless lines when the count doesn't
change.

priority_delivery_sanitizer is intended to query issuance periodically
to generate a fake sanitized response, to be put into a file that an
http server could serve as if it were the response to the investment
query.  this way, the aptera javascript that currently queries
issuance (twice for rega and regd) can be simplified to a single query
to the sanitizing server, and very little else needs to be changed.
(single query, so no need to sum, but otherwise access data via what
looks like a json query response in the same way as before.)